Revert "Make metadata reports available to local users"
