+++ seqr/views/apis/summary_data_api.py
@@ -2,8 +2,7 @@ from collections import defaultdict
 from datetime import datetime
 from django.core.exceptions import PermissionDenied
 from django.contrib.auth.models import User
-from django.contrib.postgres.aggregates import ArrayAgg
-from django.db.models import CharField, F, Q, Value
+from django.db.models import CharField, F, Value
 from django.db.models.functions import Coalesce, Concat, JSONObject, NullIf
 import json
 from random import randint
@@ -24,7 +23,7 @@ from seqr.views.utils.orm_to_json_utils import get_json_for_matchmaker_submissio
     add_individual_hpo_details, INDIVIDUAL_DISPLAY_NAME_EXPR, AIP_TAG_TYPE
 from seqr.views.utils.permissions_utils import analyst_required, user_is_analyst, get_project_guids_user_can_view, \
     login_and_policies_required, get_project_and_check_permissions, get_internal_projects
-from seqr.views.utils.anvil_metadata_utils import parse_anvil_metadata, FAMILY_ROW_TYPE, SUBJECT_ROW_TYPE, SAMPLE_ROW_TYPE, DISCOVERY_ROW_TYPE
+from seqr.views.utils.anvil_metadata_utils import parse_anvil_metadata, FAMILY_ROW_TYPE, SUBJECT_ROW_TYPE, DISCOVERY_ROW_TYPE
 from seqr.views.utils.variant_utils import get_variants_response, bulk_create_tagged_variants, DISCOVERY_CATEGORY
 from settings import SEQR_SLACK_DATA_ALERTS_NOTIFICATION_CHANNEL
 
@@ -349,122 +348,3 @@ def _get_airtable_collaborator_names(user, collaborator_ids):
         collaborator_id: collaborator_map.get(collaborator_id, {}).get('CollaboratorID')
         for collaborator_id in collaborator_ids
     }
-
-
-@login_and_policies_required
-def family_metadata(request, project_guid):
-    projects, _ = _get_metadata_projects(request, project_guid)
-
-    families_by_id = {}
-    family_individuals = defaultdict(dict)
-
-    def _add_row(row, family_id, row_type):
-        if row_type == FAMILY_ROW_TYPE:
-            families_by_id[family_id] = row
-        elif row_type == SUBJECT_ROW_TYPE:
-            family_individuals[family_id][row['participant_id']] = row
-        elif row_type == SAMPLE_ROW_TYPE:
-            family_individuals[family_id][row['participant_id']].update(row)
-        elif row_type == DISCOVERY_ROW_TYPE:
-            family = families_by_id[family_id]
-            if 'inheritance_models' not in family:
-                family.update({'genes': set(), 'inheritance_models': set()})
-            family['genes'].update({v.get('gene') or v.get('sv_name') or v.get('gene_id') or '' for v in row})
-            family['inheritance_models'].update({v['variant_inheritance'] for v in row})
-
-    parse_anvil_metadata(
-        projects, user=request.user, add_row=_add_row, omit_airtable=True, include_metadata=True, include_no_individual_families=True)
-
-    for family_id, f in families_by_id.items():
-        individuals_by_id = family_individuals[family_id]
-        proband = next((i for i in individuals_by_id.values() if i['proband_relationship'] == 'Self'), None)
-        individuals_ids = set(individuals_by_id.keys())
-        known_ids = {}
-        if proband:
-            known_ids = {
-                'proband_id': proband['participant_id'],
-                'paternal_id': proband['paternal_id'],
-                'maternal_id': proband['maternal_id'],
-            }
-            f.update(known_ids)
-            individuals_ids -= set(known_ids.values())
-
-        sorted_samples = sorted(individuals_by_id.values(), key=lambda x: x.get('date_data_generation', ''))
-        earliest_sample = next((s for s in [proband or {}] + sorted_samples if s.get('date_data_generation')), {})
-
-        inheritance_models = f.pop('inheritance_models', [])
-        f.update({
-            'individual_count': len(individuals_by_id),
-            'other_individual_ids':  '; '.join(sorted(individuals_ids)),
-            'family_structure': _get_family_structure(len(individuals_by_id), sum(1 for id in known_ids.values() if id)),
-            'data_type': earliest_sample.get('data_type'),
-            'date_data_generation': earliest_sample.get('date_data_generation'),
-            'genes': '; '.join(sorted(f.get('genes', []))),
-            'actual_inheritance': 'unknown' if inheritance_models == {'unknown'} else ';'.join(
-                sorted([i for i in inheritance_models if i != 'unknown'])),
-        })
-
-    return create_json_response({'rows': list(families_by_id.values())})
-
-
-FAMILY_STRUCTURES = {
-    1: 'singleton',
-    2: 'duo',
-    3: 'trio',
-    4: 'quad',
-}
-
-
-def _get_family_structure(num_individuals, num_known_individuals):
-    if (num_individuals and num_known_individuals == num_individuals) or (
-            num_known_individuals in {0, 3} and num_individuals == num_known_individuals + 1):
-        return FAMILY_STRUCTURES[num_individuals]
-    return 'other'
-
-
-@login_and_policies_required
-def variant_metadata(request, project_guid):
-    projects, _ = _get_metadata_projects(request, project_guid)
-
-    individuals = Individual.objects.filter(
-        family__project__in=projects, family__savedvariant__varianttag__variant_tag_type__category=DISCOVERY_CATEGORY,
-    ).distinct().annotate(
-        data_types=ArrayAgg('sample__sample_type', distinct=True, filter=Q(sample__isnull=False))
-    )
-
-    families_by_id = {}
-    participant_mme = {}
-    variant_rows = []
-
-    def _add_row(row, family_id, row_type):
-        if row_type == FAMILY_ROW_TYPE:
-            families_by_id[family_id] = row
-        elif row_type == SUBJECT_ROW_TYPE:
-            participant_mme[row['participant_id']] = row.get('MME', {})
-        elif row_type == DISCOVERY_ROW_TYPE:
-            family = families_by_id[family_id]
-            for variant in row:
-                del variant['gene_ids']
-                variant_rows.append({
-                    'MME': variant.pop('variantId') in participant_mme[variant['participant_id']].get('variant_ids', []),
-                    'phenotype_contribution': 'Full',
-                    **family,
-                    **variant,
-                })
-
-    parse_anvil_metadata(
-        projects,
-        user=request.user,
-        individual_samples={i: None for i in individuals},
-        individual_data_types={i.individual_id: i.data_types for i in individuals},
-        add_row=_add_row,
-        variant_json_fields=['clinvar', 'variantId'],
-        mme_values={'variant_ids': ArrayAgg('matchmakersubmissiongenes__saved_variant__saved_variant_json__variantId')},
-        include_metadata=True,
-        include_mondo=True,
-        omit_airtable=True,
-        proband_only_variants=True,
-        include_parent_mnvs=True,
-    )
-
-    return create_json_response({'rows': variant_rows})
