+++ ui/shared/components/table/LoadReportTable.jsx
@@ -1,9 +1,7 @@
 import React from 'react'
-import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 
-import { getUser } from 'redux/selectors'
 import { NoHoverFamilyLink } from 'shared/components/buttons/FamilyLink'
 import AwesomeBar from 'shared/components/page/AwesomeBar'
 import DataTable from 'shared/components/table/DataTable'
@@ -11,17 +9,9 @@ import { HorizontalSpacer } from 'shared/components/Spacers'
 import StateDataLoader from 'shared/components/StateDataLoader'
 import { InlineHeader, ActiveDisabledNavLink } from 'shared/components/StyledComponents'
 
-const ALL_PAGE = { downloadName: 'all_projects', path: 'all' }
-const ANALYST_VIEW_ALL_PAGES = [
-  { name: 'GREGoR', downloadName: 'all_GREGoR_projects', path: 'gregor' },
-  { name: 'Broad', ...ALL_PAGE },
-]
-const VIEW_ALL_PAGES = [{ name: 'my', ...ALL_PAGE }]
-
 const SEARCH_CATEGORIES = ['projects']
-const URL_BASE = 'summary_data'
 
-const getResultHref = urlPath => result => `/${URL_BASE}/${urlPath}/${result.key}`
+const getResultHref = urlBase => result => `/${urlBase}/${result.key}`
 
 const PROJECT_ID_FIELD = 'internal_project_id'
 
@@ -42,7 +32,7 @@ const getTableColumns = columns => ([
 ].map(({ name, ...props }) => ({ name, content: name, ...props })))
 
 const ReportTable = React.memo((
-  { projectGuid, queryForm, data, urlPath, user, columns, getColumns, idField },
+  { projectGuid, queryForm, data, urlBase, viewAllPages, columns, getColumns, idField, fileName },
 ) => (
   <div>
     <InlineHeader size="medium" content="Project:" />
@@ -50,12 +40,12 @@ const ReportTable = React.memo((
       categories={SEARCH_CATEGORIES}
       placeholder="Enter project name"
       inputwidth="350px"
-      getResultHref={getResultHref(urlPath)}
+      getResultHref={getResultHref(urlBase)}
     />
-    {(user.isAnalyst ? ANALYST_VIEW_ALL_PAGES : VIEW_ALL_PAGES).map(({ name, path }) => (
+    {viewAllPages.map(({ name, path }) => (
       <span key={path}>
         &nbsp; or &nbsp;
-        <ActiveDisabledNavLink to={`/${URL_BASE}/${urlPath}/${path}`}>{`view all ${name} projects`}</ActiveDisabledNavLink>
+        <ActiveDisabledNavLink to={`/${urlBase}/${path}`}>{`view all ${name} projects`}</ActiveDisabledNavLink>
       </span>
     ))}
     <HorizontalSpacer width={20} />
@@ -64,7 +54,7 @@ const ReportTable = React.memo((
       striped
       collapsing
       horizontalScroll
-      downloadFileName={`${ANALYST_VIEW_ALL_PAGES.find(({ path }) => path === projectGuid)?.downloadName || (data?.length && data[0][PROJECT_ID_FIELD].replace(/ /g, '_'))}_${new Date().toISOString().slice(0, 10)}_${urlPath.split('_')[0]}_metadata`}
+      downloadFileName={`${viewAllPages.find(({ path }) => path === projectGuid)?.downloadName || (data?.length && data[0][PROJECT_ID_FIELD].replace(/ /g, '_'))}_${new Date().toISOString().slice(0, 10)}_${fileName}`}
       idField={idField}
       defaultSortColumn="family_id"
       emptyContent={projectGuid ? '0 cases found' : 'Select a project to view data'}
@@ -78,20 +68,21 @@ const ReportTable = React.memo((
 ReportTable.propTypes = {
   data: PropTypes.arrayOf(PropTypes.object),
   projectGuid: PropTypes.string,
-  user: PropTypes.object,
+  viewAllPages: PropTypes.arrayOf(PropTypes.object),
   queryForm: PropTypes.node,
   columns: PropTypes.arrayOf(PropTypes.object),
   getColumns: PropTypes.func,
-  urlPath: PropTypes.string,
+  urlBase: PropTypes.string,
   idField: PropTypes.string,
+  fileName: PropTypes.string,
 }
 
 const parseResponse = ({ rows }) => ({ data: rows })
 
-const LoadReportTable = ({ match, urlPath, ...props }) => (
+const LoadReportTable = ({ match, urlBase, ...props }) => (
   <StateDataLoader
-    url={match.params.projectGuid ? `/api/${URL_BASE}/${urlPath}/${match.params.projectGuid}` : ''}
-    urlPath={urlPath}
+    url={match.params.projectGuid ? `/api/${urlBase}/${match.params.projectGuid}` : ''}
+    urlBase={urlBase}
     parseResponse={parseResponse}
     childComponent={ReportTable}
     projectGuid={match.params.projectGuid}
@@ -101,16 +92,7 @@ const LoadReportTable = ({ match, urlPath, ...props }) => (
 
 LoadReportTable.propTypes = {
   match: PropTypes.object,
-  urlPath: PropTypes.string,
+  urlBase: PropTypes.string,
 }
 
-const mapStateToProps = (state, ownProps) => {
-  const user = getUser(state)
-  return {
-    user,
-    queryFields: (user.isAnalyst && ownProps.match.params.projectGuid !== ALL_PAGE.path) ?
-      ownProps.allQueryFields : ownProps.queryFields,
-  }
-}
-
-export default connect(mapStateToProps)(LoadReportTable)
+export default LoadReportTable
